+++ src/file_dialog.rs
@@ -2743,7 +2743,15 @@ impl FileDialog {
     /// Loads the directory from the path text edit.
     fn submit_path_edit(&mut self) {
         self.close_path_edit();
-        let _ = self.load_directory(&self.canonicalize_path(&PathBuf::from(&self.path_edit_value)));
+
+        let path = self.canonicalize_path(&PathBuf::from(&self.path_edit_value));
+
+        if self.mode == DialogMode::SelectFile && path.is_file() {
+            self.state = DialogState::Selected(path);
+            return;
+        }
+
+        let _ = self.load_directory(&path);
     }
 
     /// Closes the text field at the top to edit the current path without loading